Initial Measurement refactoring (WIP)
- tof.in written to Measurement's folder instead of bin folder (but
  tof_list still uses the one in bin folder)
- Initial implementation of Serializable interface in Measurement
- Measurement's settings files no longer deleted when user checks
  'Use request settings' in Measurement settings dialog
- Added 'assert_folder_structure_equal' function to tests.utils.
  It can be used to test if folder structures match expectations.

+++ dialogs/request_settings.py
@@ -72,8 +72,8 @@ class RequestSettingsDialog(QtWidgets.QDialog):
         self.setAttribute(QtCore.Qt.WA_DeleteOnClose)
         screen_geometry = \
             QDesktopWidget.availableGeometry(QApplication.desktop())
-        self.resize(self.geometry().width() * 1.2,
-                    screen_geometry.size().height() * 0.8)
+        self.resize(int(self.geometry().width() * 1.2),
+                    int(screen_geometry.size().height() * 0.8))
 
         self.main_window = main_window
         self.request = request
@@ -200,13 +200,17 @@ class RequestSettingsDialog(QtWidgets.QDialog):
             default_measurement_settings_file = Path(
                 self.request.default_measurement.directory,
                 "Default.measurement")
+
             self.request.default_measurement.measurement_to_file(
                 default_measurement_settings_file)
+
             self.request.default_measurement.profile_to_file(Path(
                 self.request.default_measurement.directory,
                 "Default.profile"))
+
             self.request.default_measurement.run.to_file(
                 default_measurement_settings_file)
+            
             self.request.default_target.to_file(
                 None, default_measurement_settings_file)
